made class an inner class
